Move payload module into analyze module
