+++ mythx_cli/analyze/payload/bytecode.py
