+++ mythx_cli/analyze/payload/util.py
